Don't use DST offset - fixes 1h time shift in round trip
The current implmentation calculates a an offset based on Daylight
Saving Time being in effect at the Epoch:

  t -. offset +. (if tm.Unix.tm_isdst then 3600. else 0.)

This leads to suprises in the Europe/London timezone where unparsing and
parsing a date leads to an offset - see this issue:

  https://github.com/ocaml-community/ISO8601.ml/issues/17

The following change eliminates this and makes all unit tests pass.
Somewhat surprisingly the knowledge about this offset is encoded both in
the implementation and the unit tests.

Likewise surprising, the offset is not zero in the Europe/London
timezone but it seems to be consistent over various Unix
implementations.

Signed-off-by: Christian Lindig <lindig@gmail.com>